Decoupling FfmpegMusicService into several interfaces and implementations
