+++ MagicConchBot/Common/Enums/AudioState.cs
@@ -1,0 +1,10 @@
+namespace MagicConchBot.Common.Enums
+{
+    public enum AudioState
+    {
+        Stopped = 0,
+        Loading,
+        Paused,
+        Playing
+    }
+}
