--- MagicConchBot/Common/Enums/MusicState.cs
@@ -1,10 +1,0 @@
-namespace MagicConchBot.Common.Enums
-{
-    public enum MusicState
-    {
-        Stopped = 0,
-        Loading,
-        Paused,
-        Playing
-    }
-}
