+++ MagicConchBot/Common/Interfaces/ISongPlayer.cs
@@ -1,0 +1,26 @@
+﻿using System.Threading.Tasks;
+using Discord.Audio;
+using MagicConchBot.Common.Enums;
+using MagicConchBot.Common.Types;
+
+namespace MagicConchBot.Common.Interfaces
+{
+    public interface ISongPlayer
+    {
+        float Volume { get; set; }
+        AudioState AudioState { get; }
+        Task PlaySong(IAudioClient client, Song song);
+        void Stop();
+        void Pause();
+    }
+
+    public interface IFileProvider
+    {
+        Task<string> GetStreamingFile(Song song);
+    }
+
+    public interface ISongResolver
+    {
+        Task<string> GetSongStreamUrl(string url);
+    }
+}
