+++ MagicConchBot/Common/Types/PomfResult.cs
@@ -1,18 +1,29 @@
 ﻿using System.Collections.Generic;
+using Newtonsoft.Json;
 
 namespace MagicConchBot.Common.Types
 {
     public class PomfFile
     {
-        public string name { get; set; }
-        public string url { get; set; }
-        public string fullurl { get; set; }
-        public int size { get; set; }
+        [JsonProperty("name")]
+        public string Name { get; set; }
+
+        [JsonProperty("url")]
+        public string Url { get; set; }
+
+        [JsonProperty("fullurl")]
+        public string FullUrl { get; set; }
+
+        [JsonProperty("size")]
+        public int Size { get; set; }
     }
 
     public class PomfResult
     {
-        public bool success { get; set; }
-        public List<PomfFile> files { get; set; }
+        [JsonProperty("success")]
+        public bool Success { get; set; }
+
+        [JsonProperty("files")]
+        public List<PomfFile> Files { get; set; }
     }
 }
