+++ MagicConchBot/run.bat
@@ -1,0 +1,1 @@
+C:\Cygwin64\bin\bash.exe -l '/cygdrive/c/users/shred/onedrive/documents/visual studio 2017/projects/magicconchbot/magicconchbot/run.sh'
