+++ MagicConchBot/run.sh
@@ -1,0 +1,1 @@
+ssh -t bot './kill.sh ; sleep 1 ; ./run.sh'
